+++ include/internal/quic_ssl.h
@@ -98,6 +98,10 @@ BIO *ossl_quic_conn_get_net_rbio(const SSL *s);
 BIO *ossl_quic_conn_get_net_wbio(const SSL *s);
 __owur int ossl_quic_conn_set_initial_peer_addr(SSL *s,
     const BIO_ADDR *peer_addr);
+__owur int ossl_quic_conn_set_size_probes(SSL *s, const uint16_t *sizes,
+    size_t n);
+__owur int ossl_quic_conn_get_size_probes(SSL *s, uint16_t *confirmed,
+    uint64_t *acked);
 __owur SSL *ossl_quic_conn_stream_new(SSL *s, uint64_t flags);
 __owur SSL *ossl_quic_get0_connection(SSL *s);
 __owur SSL *ossl_quic_get0_listener(SSL *s);